Refactor common Checkbox code into modules
